chore: extract utils folder (#876)
