+++ src/DEPS.list
@@ -1,6 +1,7 @@
 [*]
 ./tools/
 ./mcp/
+./utils/
 
 [program.ts]
 ***
