+++ src/extension/DEPS.list
@@ -1,3 +1,3 @@
 [*]
-../
 ../mcp/
+../utils/
