+++ src/loopTools/DEPS.list
@@ -3,3 +3,4 @@
 ../loop/
 ../mcp/
 ../tools/
+../utils/
