+++ src/mcp/DEPS.list
@@ -1,7 +1,2 @@
 [*]
-../log.js
-../manualPromise.js
-../httpServer.js
-
-[proxyBackend.ts]
-../package.js
+../utils/
