+++ tests/roots.spec.ts
@@ -19,7 +19,7 @@ import path from 'path';
 import { pathToFileURL } from 'url';
 
 import { test, expect } from './fixtures.js';
-import { createHash } from '../src/utils.js';
+import { createHash } from '../src/utils/guid.js';
 
 const p = process.platform === 'win32' ? 'c:\\non\\existent\\folder' : '/non/existent/folder';
